test: split tap AUT frame spec by command domain
Extract extractDom, extractAria, and extractInspect coverage into
dom.spec.ts, aria.spec.ts, and inspect.spec.ts. frame.spec.ts now
covers only the aut/frame.ts helpers (resolveAutFrame, parsePositiveInt).
